requirements.txt
dymo-bluetooth
python-barcode
flask[async]